+++ hamza-client/src/store/wishlist/wishlist-store.tsx
@@ -1,12 +1,5 @@
-import React, { useEffect, useState } from 'react';
-import { useRegion, useMeCustomer } from 'medusa-react';
-import { useQuery, useMutation } from '@tanstack/react-query';
-import axios from 'axios';
 import { create } from 'zustand';
 import { persist, createJSONStorage } from 'zustand/middleware';
-import { useCustomerAuthStore } from '@store/customer-auth/customer-auth';
-
-// TODO: Refactor to pull useWishlistStore from this component or the Wishlist component to create a sort of separation (e.g. useWishlistStore.tsx)
 
 type WishlistItem = {
     id: string;
@@ -60,43 +53,3 @@ const useWishlistStore = create<WishlistType>()(
 );
 
 export default useWishlistStore;
-
-export const Wishlist = () => {
-    const { customer_id } = useCustomerAuthStore((state) => ({
-        customer_id: state.customer_id,
-    }));
-
-    const addWishlistMutation = useMutation(
-        (product_id) =>
-            axios.post(`localhost:9000/custom/wishlist/item`, {
-                customer_id: customer_id,
-                product_id: product_id,
-            }),
-        {
-            onSuccess: (data) => {
-                console.log('Adding Wish list item in DB!');
-            },
-            onError: (error) => {
-                console.log('Error adding item to wishlist', error);
-            },
-        }
-    );
-
-    // now we will refactor removeWishItem to removeWishlistItemMutation
-
-    const removeWishlistItemMutation = useMutation(
-        (product_id) =>
-            axios.delete(`localhost:9000/custom/wishlist/item`, {
-                customer_id: customer_id,
-                product_id: product_id,
-            }),
-        {
-            onSuccess: (data) => {
-                console.log('Removing Wish List item in DB');
-            },
-            onError: (error) => {
-                console.log('Error removing item from wishlist', error);
-            },
-        }
-    );
-};
